# docs/swagger.yaml
basePath: /
definitions:
  http.externalApplicationRequest:
    properties:
      accountHolder:
        maxLength: 150
        minLength: 1
        type: string
      birthDate:
        type: string
      companyName:
        maxLength: 150
        minLength: 1
        type: string
      consumptionForecast:
        minimum: 0
        type: integer
      consumptionPreviousYear:
        minimum: 0
        type: integer
      electricHotWater:
        type: boolean
      electricVehicle:
        type: boolean
      email:
        type: string
      feedInForecast:
        minimum: 0
        type: integer
      firstname:
        maxLength: 100
        minLength: 1
        type: string
      heatPump:
        type: boolean
      iban:
        maxLength: 50
        minLength: 15
        type: string
      lastname:
        maxLength: 100
        minLength: 1
        type: string
      memberType:
        enum:
        - private
        - farmer
        - municipality
        - company
        - association
        type: string
      membershipStartDate:
        description: Configurable fields (PROJ-8)
        type: string
      meteringPoints:
        items:
          $ref: '#/definitions/shared.CreateMeteringPointRequest'
        maxItems: 10
        minItems: 1
        type: array
      personsInHousehold:
        minimum: 0
        type: integer
      phone:
        maxLength: 50
        type: string
      privacyAccepted:
        type: boolean
      pvPowerKwp:
        minimum: 0
        type: number
      registerNumber:
        maxLength: 50
        type: string
      residentCity:
        maxLength: 255
        minLength: 1
        type: string
      residentStreet:
        maxLength: 255
        minLength: 1
        type: string
      residentStreetNumber:
        maxLength: 50
        minLength: 1
        type: string
      residentZip:
        maxLength: 20
        minLength: 1
        type: string
      sepaMandateAccepted:
        type: boolean
      titel:
        maxLength: 50
        type: string
      uidNumber:
        maxLength: 50
        type: string
    required:
    - accountHolder
    - email
    - iban
    - memberType
    - meteringPoints
    - privacyAccepted
    - residentCity
    - residentStreet
    - residentStreetNumber
    - residentZip
    - sepaMandateAccepted
    type: object
...
  shared.ApplicationResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      referenceNumber:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  shared.ApplicationStatus:
    enum:
    - draft
    - submitted
    - under_review
    - needs_info
    - approved
    - rejected
    - imported
    - import_failed
    type: string
    x-enum-varnames:
    - StatusDraft
    - StatusSubmitted
    - StatusUnderReview
    - StatusNeedsInfo
    - StatusApproved
    - StatusRejected
    - StatusImported
    - StatusImportFailed
...
  shared.ConsentInput:
    properties:
      isCentralPolicy:
        type: boolean
      title:
        type: string
      url:
        type: string
    type: object
  shared.CreateApplicationRequest:
    properties:
      accountHolder:
        maxLength: 150
        minLength: 1
        type: string
      accuracyConfirmed:
        type: boolean
      birthDate:
        type: string
      companyName:
        maxLength: 150
        minLength: 1
        type: string
      consumptionForecast:
        minimum: 0
        type: integer
      consumptionPreviousYear:
        minimum: 0
        type: integer
      electricHotWater:
        type: boolean
      electricVehicle:
        type: boolean
      email:
        type: string
      feedInForecast:
        minimum: 0
        type: integer
      firstname:
        maxLength: 100
        minLength: 1
        type: string
      heatPump:
        type: boolean
      iban:
        maxLength: 50
        minLength: 15
        type: string
      lastname:
        maxLength: 100
        minLength: 1
        type: string
      memberType:
        enum:
        - private
        - farmer
        - municipality
        - company
        - association
        type: string
      membershipStartDate:
        description: Configurable application-level fields (PROJ-8)
        type: string
      meteringPoints:
        items:
          $ref: '#/definitions/shared.CreateMeteringPointRequest'
        maxItems: 10
        minItems: 1
        type: array
      personsInHousehold:
        minimum: 0
        type: integer
      phone:
        maxLength: 50
        type: string
      privacyAccepted:
        type: boolean
      privacyVersion:
        type: string
      pvPowerKwp:
        minimum: 0
        type: number
      rcNumber:
        type: string
      registerNumber:
        maxLength: 50
        type: string
      residentCity:
        maxLength: 100
        minLength: 1
        type: string
      residentStreet:
        maxLength: 100
        minLength: 1
        type: string
      residentStreetNumber:
        maxLength: 50
        minLength: 1
        type: string
      residentZip:
        maxLength: 20
        minLength: 1
        type: string
      sepaMandateAccepted:
        type: boolean
      titel:
        maxLength: 50
        type: string
      turnstileToken:
        description: Cloudflare Turnstile token (PROJ-16) — optional, verified server-side
          when TURNSTILE_SECRET_KEY is set
        type: string
      uidNumber:
        maxLength: 50
        type: string
    required:
    - accountHolder
    - accuracyConfirmed
    - email
    - iban
    - memberType
    - meteringPoints
    - privacyAccepted
    - privacyVersion
    - rcNumber
    - residentCity
    - residentStreet
    - residentStreetNumber
    - residentZip
    - sepaMandateAccepted
    type: object
  shared.CreateMeteringPointRequest:
    properties:
      direction:
        enum:
        - CONSUMPTION
        - PRODUCTION
        type: string
      installationName:
        maxLength: 100
        type: string
      installationNumber:
        maxLength: 50
        type: string
      meteringPoint:
        maxLength: 33
        type: string
      participationFactor:
        maximum: 100
        minimum: 1
        type: integer
      transformer:
        description: Configurable metering-point-level fields (PROJ-8)
        maxLength: 100
        type: string
    required:
    - direction
    - meteringPoint
    - participationFactor
    type: object
...
  shared.ErrorResponse:
    properties:
      code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  shared.SubmitRequest:
    properties:
      consents:
        items:
          $ref: '#/definitions/shared.ConsentInput'
        type: array
    type: object
  shared.SubmitResponse:
    properties:
      id:
        type: string
      referenceNumber:
        type: string
      status:
        $ref: '#/definitions/shared.ApplicationStatus'
      submittedAt:
        type: string
    type: object
  shared.UpdateApplicationRequest:
    properties:
      accountHolder:
        maxLength: 150
        minLength: 1
        type: string
      accuracyConfirmed:
        type: boolean
      birthDate:
        type: string
      companyName:
        maxLength: 150
        minLength: 1
        type: string
      email:
        type: string
      firstname:
        maxLength: 100
        minLength: 1
        type: string
      iban:
        maxLength: 50
        minLength: 15
        type: string
      lastname:
        maxLength: 100
        minLength: 1
        type: string
      memberType:
        enum:
        - private
        - farmer
        - municipality
        - company
        - association
        type: string
      meteringPoints:
        items:
          $ref: '#/definitions/shared.CreateMeteringPointRequest'
        maxItems: 10
        minItems: 1
        type: array
      phone:
        maxLength: 50
        type: string
      privacyAccepted:
        type: boolean
      privacyVersion:
        type: string
      registerNumber:
        maxLength: 50
        type: string
      residentCity:
        maxLength: 100
        minLength: 1
        type: string
      residentStreet:
        maxLength: 100
        minLength: 1
        type: string
      residentStreetNumber:
        maxLength: 50
        minLength: 1
        type: string
      residentZip:
        maxLength: 20
        minLength: 1
        type: string
      sepaMandateAccepted:
        type: boolean
      titel:
        maxLength: 50
        type: string
      uidNumber:
        maxLength: 50
        type: string
    type: object
host: member-onboarding.eegfaktura.at
info:
  contact: {}
  description: |-
    REST API for EEG (Energiegemeinschaft) member self-service registration and admin review.

    **Auth schemes:**
    - **Public** endpoints (`/api/public/*`): no auth required. Rate-limited + optional Turnstile CAPTCHA.
    - **Admin** endpoints (`/api/admin/*`): Keycloak Bearer JWT (`Authorization: Bearer <token>`).
    - **External** endpoints (`/api/external/*`): API key (`Authorization: Bearer moak_<key>`).
  title: eegfaktura Member Onboarding API
  version: "1.0"
paths:
...
  /api/external/v1/applications:
    post:
      consumes:
      - application/json
      description: Creates and immediately submits a member application in a single
        request. Intended for ERP/third-party integrations. The EEG RC number is derived
        from the API key — no rc_number field in the body.
      parameters:
      - description: Member application data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.externalApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: id and referenceNumber of the created application
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: RC number not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "409":
          description: Duplicate metering point
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "410":
          description: Registration deactivated
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Submit application (external)
      tags:
      - External
  /api/public/applications:
    post:
      consumes:
      - application/json
      description: Creates a new member application in status `draft`. Rate-limited
        (10 req/10 min per IP). Protected by Cloudflare Turnstile when configured.
      parameters:
      - description: Application data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/shared.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/shared.ApplicationResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "409":
          description: Duplicate metering point
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "410":
          description: Registration deactivated
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "422":
          description: Turnstile verification failed
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Create application draft
      tags:
      - Public
  /api/public/applications/{id}:
    put:
      consumes:
      - application/json
      description: Updates a member application in status `draft`. Only drafts can
        be updated; submitted applications return 409.
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated application data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/shared.UpdateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.ApplicationResponse'
        "400":
          description: Validation error or invalid UUID
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "409":
          description: Application already submitted
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Update application draft
      tags:
      - Public
  /api/public/applications/{id}/submit:
    post:
      consumes:
      - application/json
      description: Transitions the application from `draft` to `submitted`. Triggers
        confirmation email to the member. Consents for legal documents can be passed
        in the request body.
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      - description: Optional legal document consents
        in: body
        name: body
        schema:
          $ref: '#/definitions/shared.SubmitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.SubmitResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Application not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "409":
          description: Invalid status transition
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      summary: Submit application
      tags:
      - Public
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'EEG API key. Format: "Bearer moak_<key>"'
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: 'Keycloak JWT. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"